+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 
 
 
+### [06/04/2025](https://github.com/vungnv10103/ServerAvatar258)
+
+- [avatar/00:10](https://github.com/vungnv10103/ServerAvatar258/issues) update `save captcha`, update `index.html`.
+
+
 ### [01/04/2025](https://github.com/vungnv10103/ServerAvatar258)
 
 - [avatar/00:25](https://github.com/vungnv10103/ServerAvatar258/issues) add `module train captcha`, add `template captcha`.